% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleAdd.R
\name{roleAdd_isNew}
\alias{roleAdd_isNew}
\title{检测角色是否存在}
\usage{
roleAdd_isNew(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  role_name = "standard"
)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{role_name}{角色名}
}
\value{
返回值
}
\description{
检测角色是否存在
}
\examples{
roleAdd_isNew()
}
