% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleAdd.R
\name{roleAdd_query}
\alias{roleAdd_query}
\title{检测角色是否存在}
\usage{
roleAdd_query(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  role_name = "standard"
)

roleAdd_query(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  role_name = "standard"
)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{role_name}{角色名}
}
\value{
返回值

返回值
}
\description{
检测角色是否存在

检测角色是否存在
}
\examples{
roleAdd_isNew()
roleAdd_isNew()
}
