% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleAdd.R
\name{roleAdd_save}
\alias{roleAdd_save}
\title{检测角色是否存在}
\usage{
roleAdd_save(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  role_name = "demo",
  note = ""
)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{role_name}{角色名}

\item{note}{备注}
}
\value{
返回值
}
\description{
检测角色是否存在
}
\examples{

roleAdd_save()
}
