% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleAdd.R
\name{roleAddServer}
\alias{roleAddServer}
\title{角色新增的逻辑处理}
\usage{
roleAddServer(input, output, session, app_id)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{app_id}{程序}
}
\value{
返回值
}
\description{
角色新增的逻辑处理
}
\examples{
roleAddServer()
}
