% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleQuery.R
\name{roleQuery_enable}
\alias{roleQuery_enable}
\title{用户反禁用}
\usage{
roleQuery_enable(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  role_name = ""
)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{role_name}{用户名}
}
\value{
返回值
}
\description{
用户反禁用
}
\examples{
roleQuery_query()
}
