% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleQuery.R
\name{roleQuery_query}
\alias{roleQuery_query}
\title{用户查询}
\usage{
roleQuery_query(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  role_name = "",
  status = -1
)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{role_name}{用户名}

\item{status}{禁用状态}
}
\value{
返回值
}
\description{
用户查询
}
\examples{
roleQuery_query()
}
