% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleRight.R
\name{roleRight_object_query}
\alias{roleRight_object_query}
\title{查询用户}
\usage{
roleRight_object_query(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  role_name = "财务总监",
  type = "system"
)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{role_name}{角色}

\item{type}{类型}
}
\value{
返回值
}
\description{
查询用户
}
\examples{
roleRight_object_query()
}
