% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleRight.R
\name{roleRight_object_update}
\alias{roleRight_object_update}
\title{设置权限}
\usage{
roleRight_object_update(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  role_name = "财务总监",
  type = "system",
  system_id = "A",
  module_id = "",
  show = 1
)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{role_name}{角色}

\item{type}{类型}

\item{system_id}{系统ID}

\item{module_id}{功能ID}

\item{show}{是否授权}
}
\value{
返回值
}
\description{
设置权限
}
\examples{
roleRight_object_update()
}
