% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roleRight.R
\name{roleRightServer}
\alias{roleRightServer}
\title{用户新增的逻辑处理}
\usage{
roleRightServer(input, output, session, app_id)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{app_id}{程序}
}
\value{
返回值
}
\description{
用户新增的逻辑处理
}
\examples{
roleRightServer()
}
