% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userAdd.R
\name{userAdd_query}
\alias{userAdd_query}
\title{检测用户是否存在}
\usage{
userAdd_query(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  user_name = "demo"
)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{user_name}{用户名}
}
\value{
返回值
}
\description{
检测用户是否存在
}
\examples{
userAdd_isNew()
}
