% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userQuery.R
\name{userQuery_enable}
\alias{userQuery_enable}
\title{用户反禁用}
\usage{
userQuery_enable(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  user_name = ""
)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{user_name}{用户名}
}
\value{
返回值
}
\description{
用户反禁用
}
\examples{
userQuery_query()
}
