% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userRoleSetting.R
\name{userRoleSetting_role_update}
\alias{userRoleSetting_role_update}
\title{查询用户}
\usage{
userRoleSetting_role_update(
  conn = tsda::conn_rds("rdbe"),
  app_id = "cpdms",
  user_name = "韩向松",
  role_name = "admin"
)
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}

\item{user_name}{用户名}

\item{role_name}{密码}
}
\value{
返回值
}
\description{
查询用户
}
\examples{
userRoleSetting_user_query()
}
