% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userRoleSetting.R
\name{userRoleSetting_user_query}
\alias{userRoleSetting_user_query}
\title{查询用户}
\usage{
userRoleSetting_user_query(conn = tsda::conn_rds("rdbe"), app_id = "cpdms")
}
\arguments{
\item{conn}{连接}

\item{app_id}{程序ID}
}
\value{
返回值
}
\description{
查询用户
}
\examples{
userRoleSetting_user_query()
}
